move common RNN utilities into src/Conversion/ONNXConversionCommon
