+++ src/Conversion/CMakeLists.txt
@@ -1,5 +1,6 @@
 # SPDX-License-Identifier: Apache-2.0
 
+add_subdirectory(ONNXConversionCommon)
 add_subdirectory(ONNXToKrnl)
 add_subdirectory(KrnlToLLVM)
 add_subdirectory(KrnlToAffine)
