+++ src/Conversion/ONNXConversionCommon/CMakeLists.txt
@@ -1,0 +1,12 @@
+# SPDX-License-Identifier: Apache-2.0
+
+add_onnx_mlir_library(OMONNXConversionCommon
+  RNN/RNNBase.cpp
+  RNN/LSTM.cpp
+
+  LINK_LIBS PUBLIC
+
+  OMONNXOps
+  OMSupport
+  MLIRTransforms
+)
